pyproject.toml
[project]
name = "promptdev"
version = "0.1.0"
description = "Python-native prompt evaluation tool using PydanticAI"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic-ai==1.0.0b1",
    "pydantic-evals",
    "pydantic>=2.0.0",
    "click>=8.0.0",
    "pyyaml>=6.0.0",
    "rich>=13.0.0",
    "httpx>=0.27.0",
    "jinja2>=3.0.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0", "pytest-asyncio>=0.21.0", "ruff>=0.1.0"]
logfire = ["logfire>=0.1.0"]

[project.scripts]
promptdev = "promptdev.cli:cli"

[build-system] # TODO use uv build instead
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py312"
extend-exclude = ["__pycache__", ".venv", "build", "dist"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG", # flake8-unused-arguments
    "SIM", # flake8-simplify
]
ignore = [
    "E501",   # line too long, handled by formatter
    "B008",   # do not perform function calls in argument defaults
    "ARG001", # unused function argument
    "ARG002", # unused method argument
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.logfire]
ignore_no_config = true
